requirements.txt
numpy==1.17.2
opencv-python==4.1.1.26
Keras==2.3.1
Pillow==6.2.0
tensorflow==2.0.0